fix: test by extracted dummy content component
