+++ apps/web/src/components/templates/in-app-editor/InAppEditorBlock.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { InAppWidgetPreview } from '../../widget/InAppWidgetPreview';
-import { ContentContainer } from './ContentContainer';
+import { ContentContainer } from './content/ContentContainer';
 import { Control, Controller } from 'react-hook-form';
 import { IForm } from '../use-template-controller.hook';
